move nozzle manager
