+++ api/src/opentrons/hardware_control/dev_types.py
@@ -28,7 +28,7 @@ from opentrons_shared_data.gripper import (
 from opentrons.drivers.types import MoveSplit
 from opentrons.types import Mount
 from opentrons.hardware_control.types import GripperJawState
-from opentrons.hardware_control.instruments.nozzle_manager import NozzleMap
+from opentrons.hardware_control.nozzle_manager import NozzleMap
 
 
 class InstrumentSpec(TypedDict):
@@ -96,6 +96,7 @@ class PipetteDict(InstrumentDict):
     default_push_out_volume: Optional[float]
     supported_tips: Dict[PipetteTipType, SupportedTipsDefinition]
     current_nozzle_map: Optional[NozzleMap]
+
 
 class PipetteStateDict(TypedDict):
     tip_detected: bool
