+++ api/src/opentrons/hardware_control/instruments/ot3/pipette.py
@@ -45,7 +45,7 @@ from opentrons_shared_data.pipette import (
 )
 from opentrons.hardware_control.types import CriticalPoint, OT3Mount
 from opentrons.hardware_control.errors import InvalidMoveError
-from opentrons.hardware_control.instruments import nozzle_manager
+from opentrons.hardware_control import nozzle_manager
 
 mod_log = logging.getLogger(__name__)
