+++ api/src/opentrons/protocol_engine/resources/pipette_data_provider.py
@@ -12,7 +12,7 @@ from opentrons_shared_data.pipette import (
 
 
 from opentrons.hardware_control.dev_types import PipetteDict
-from opentrons.hardware_control.instruments.nozzle_manager import (
+from opentrons.hardware_control.nozzle_manager import (
     NozzleConfigurationManager,
     NozzleMap,
 )
